msm: msm_bus: Fix the type error affecting sum of abs
On legacy chipsets, uint16 was being used to write bandwidth
values to the Fabric registers. However, NoCs support 64-bit integers.
The issue surfaces only when sum of active ab values exceeds the max
ib value in the system, as then the sum of ab values don't impact the
clock anymore. This is exactly what was observed in H-flip video usecase.

This patch fixex the bandwidth sum type due to which clocks were not
updated in those cases where requested ab values are higher than the ib
values.

CRs-Fixed: 480365

+++ arch/arm/mach-msm/msm_bus/msm_bus_arb.c
@@ -314,7 +314,7 @@ static int update_path(int curr, int pnode, uint64_t req_clk, uint64_t req_bw,
 	struct msm_bus_inode_info *info;
 	int next_pnode;
 	int64_t add_bw = req_bw - curr_bw;
-	unsigned bwsum = 0;
+	uint64_t bwsum = 0;
 	uint64_t req_clk_hz, curr_clk_hz, bwsum_hz;
 	int *master_tiers;
 	struct msm_bus_fabric_device *fabdev = msm_bus_get_fabric_device
@@ -398,7 +398,7 @@ static int update_path(int curr, int pnode, uint64_t req_clk, uint64_t req_bw,
 		/* Update Bandwidth */
 		fabdev->algo->update_bw(fabdev, hop, info, add_bw,
 			master_tiers, ctx);
-		bwsum = (uint16_t)*hop->link_info.sel_bw;
+		bwsum = *hop->link_info.sel_bw;
 		/* Update Fabric clocks */
 		curr_clk_hz = BW_TO_CLK_FREQ_HZ(hop->node_info->buswidth,
 			curr_clk);
